src/main/java/roomescape/auth/service/AuthService.java
package roomescape.auth.service;

import org.springframework.stereotype.Service;
import roomescape.auth.jwt.JwtTokenProvider;
import roomescape.auth.dto.LoginRequest;
import roomescape.member.domain.Member;
import roomescape.auth.dto.LoginMember;
import roomescape.member.dto.MemberResponse;
import roomescape.member.repository.MemberRepository;

@Service
public class AuthService {

    private final MemberRepository memberRepository;
    private final JwtTokenProvider jwtTokenProvider;

    public AuthService(MemberRepository memberRepository, JwtTokenProvider jwtTokenProvider) {
        this.memberRepository = memberRepository;
        this.jwtTokenProvider = jwtTokenProvider;
    }

    public String createMemberToken(LoginRequest loginRequest) {
        Member member = findByEmailAndPassword(loginRequest);

        return jwtTokenProvider.generateToken(member);
    }

    public Member findByEmailAndPassword(LoginRequest loginRequest) {
        return memberRepository.findByEmailAndPassword(loginRequest.email(), loginRequest.password())
                .orElseThrow(() -> new IllegalArgumentException("일치하지 않는 이메일 또는 비밀번호입니다."));
    }

    public MemberResponse findMemberNameByLoginMember(LoginMember loginMember) {
        Member member = memberRepository.findById(loginMember.id())
                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 회원입니다."));

        return MemberResponse.toResponse(member);
    }
}
